input parsing, refactoring, comments
